Handle Python 3 iterating over modified dictionaries
Under Python 3, the dictionary values() call gives a view
object that raises RuntimeError if the dictionary is modified
while iterating. Technically that also happened under Python 2
if the dictionary itervalues() method was used.

This uses a simple _copiedvalues() function to get a standalone
list of values efficiently on either version of Python. The
recursion functions for find, save/write, and close use this
to prevent exceptions. The behavior under Python 2 is unchanged.